refactor: move brc custom workflow page to explicit assemblies route (#1637)
