--- pages/data/[entityListType]/[entityId]/analyze/custom.tsx
@@ -1,84 +1,0 @@
-import { getPageMeta } from "@/common/meta/utils";
-import { config } from "@/config/config";
-import { EntityDataGate } from "@repo/shared/components/EntityDataGate/entityDataGate";
-import { getEntities } from "@repo/shared/services/staticGeneration/entities/utils";
-import { seedDatabase } from "@repo/shared/utils/seedDatabase/utils";
-import { WorkflowInputsView } from "@repo/shared/views/WorkflowInputsView/workflowInputsView";
-import { CUSTOM_WORKFLOW } from "@repo/shared/workflow/custom";
-import {
-  type GetStaticPaths,
-  type GetStaticPathsResult,
-  type GetStaticProps,
-  type GetStaticPropsContext,
-} from "next";
-import { type ParsedUrlQuery } from "querystring";
-import { type JSX } from "react";
-
-interface Params extends ParsedUrlQuery {
-  entityId: string;
-  entityListType: string;
-}
-
-export interface Props {
-  entityId: string;
-  pageDescription?: string;
-  pageTitle?: string;
-  trsId: string;
-}
-
-export const getStaticPaths: GetStaticPaths<Params> = async () => {
-  const paths: GetStaticPathsResult<Params>["paths"] = [];
-
-  for (const entityConfig of config().entities) {
-    const { route: entityListType } = entityConfig;
-
-    // Only statically generate paths for each assembly.
-    if (entityListType !== "assemblies") continue;
-
-    await seedDatabase(entityListType, entityConfig);
-
-    const entities = await getEntities(entityConfig);
-
-    for (const entity of entities.hits) {
-      const entityId = entityConfig.getId?.(entity);
-
-      if (!entityId) continue;
-
-      paths.push({ params: { entityId, entityListType } });
-    }
-  }
-
-  return { fallback: false, paths };
-};
-
-export const getStaticProps: GetStaticProps<Props> = async ({
-  params,
-}: GetStaticPropsContext) => {
-  const { entityId } = params as Params;
-
-  if (!entityId) return { notFound: true };
-
-  // Pass the custom workflow TRS ID as a prop to the page, so that the correct workflow is loaded in the WorkflowInputsView.
-  return {
-    props: {
-      entityId,
-      ...getPageMeta(config().appKey).CUSTOM_WORKFLOW,
-      trsId: CUSTOM_WORKFLOW.trsId,
-    },
-  };
-};
-
-/**
- * Custom workflow view page.
- * @param props - Page props.
- * @returns Custom workflow view component.
- */
-const Page = (props: Props): JSX.Element => {
-  return (
-    <EntityDataGate>
-      <WorkflowInputsView {...props} />
-    </EntityDataGate>
-  );
-};
-
-export default Page;
